+++ src/constants/poolAddress.ts
@@ -4,7 +4,7 @@ export const POOL_ADDRESSES: Record<
   SupportPoolCoins,
   {
     lendingPoolAddress: string;
-    collateralPoolAddress: string;
+    collateralPoolAddress?: string; // not all pool has collateral
   }
 > = {
   usdc: {
@@ -90,5 +90,13 @@ export const POOL_ADDRESSES: Record<
       '0x6fc7d4211fc7018b6c75e7b908b88f2e0536443239804a3d32af547637bd28d7',
     collateralPoolAddress:
       '0xff677a5d9e9dc8f08f0a8681ebfc7481d1c7d57bc441f2881974adcdd7b13c31',
+  },
+  fud: {
+    lendingPoolAddress:
+      '0x14367ddca30e2860cb89ed4eaca20c7ece260c5d59dd9990d2c85a8321326acb',
+  },
+  deep: {
+    lendingPoolAddress:
+      '0xf4a67ffb43da1e1c61c049f188f19463ea8dbbf2d5ef4722d6df854ff1b1cc03',
   },
 };